requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==2.11.1
attrs==23.1.0
certifi==2023.7.22
et-xmlfile==1.1.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pluggy==1.3.0
PySocks==1.7.1
pytest==7.4.0
pytest-rerunfailures==12.0
selenium==4.11.2
shutup==0.2.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
wsproto==1.2.0
